Fixes wrong WKT conversion at LOI creation time + Moved code into lib (#2057)
