+++ functions/src/import-geojson.ts
@@ -22,8 +22,8 @@ import JSONStream from 'jsonstream-ts';
 import {canImport} from './common/auth';
 import {DecodedIdToken} from 'firebase-admin/auth';
 import {GroundProtos} from '@ground/proto';
-import {toDocumentData} from '@ground/lib';
-import {Feature, GeoJsonProperties, Geometry, Position} from 'geojson';
+import {toDocumentData, toGeometryPb} from '@ground/lib';
+import {Feature, GeoJsonProperties} from 'geojson';
 import {ErrorHandler} from './handlers';
 
 import Pb = GroundProtos.ground.v1beta1;
@@ -169,7 +169,7 @@ function toLoiPb(
   });
 }
 
-function toLoiPbProperties(properties: GeoJsonProperties): {
+export function toLoiPbProperties(properties: GeoJsonProperties): {
   [k: string]: Pb.LocationOfInterest.Property;
 } {
   return Object.fromEntries(
@@ -183,58 +183,4 @@ function toLoiPbProperty(value: any): Pb.LocationOfInterest.Property {
       ? {numericValue: value}
       : {stringValue: value?.toString() || ''}
   );
-}
-
-function toGeometryPb(geometry: Geometry): Pb.Geometry {
-  switch (geometry.type) {
-    case 'Point':
-      return toPointGeometry(geometry.coordinates);
-    case 'Polygon':
-      return toPolygonGeometry(geometry.coordinates);
-    case 'MultiPolygon':
-      return toMultiPolygonGeometry(geometry.coordinates);
-    default:
-      throw new Error(`Unsupported GeoJSON type '${geometry.type}'`);
-  }
-}
-
-function toPointGeometry(position: Position): Pb.Geometry {
-  const coordinates = toCoordinatesPb(position);
-  const point = new Pb.Point({coordinates});
-  return new Pb.Geometry({point});
-}
-
-function toCoordinatesPb(position: Position): Pb.Coordinates {
-  const [longitude, latitude] = position;
-  if (longitude === undefined || latitude === undefined)
-    throw new Error('Missing coordinate(s)');
-  return new Pb.Coordinates({longitude, latitude});
-}
-
-function toPolygonPb(positions: Position[][]): Pb.Polygon {
-  const [shellCoords, ...holeCoords] = positions;
-  // Ignore if shell is missing.
-  if (!shellCoords)
-    throw new Error('Missing required polygon shell coordinates');
-  const shell = toLinearRingPb(shellCoords);
-  const holes = holeCoords?.map(h => toLinearRingPb(h));
-  return new Pb.Polygon({shell, holes});
-}
-
-function toPolygonGeometry(positions: Position[][]): Pb.Geometry {
-  const polygon = toPolygonPb(positions);
-  return new Pb.Geometry({polygon});
-}
-
-function toLinearRingPb(positions: Position[]): Pb.LinearRing {
-  const coordinates = positions.map(p => toCoordinatesPb(p));
-  return new Pb.LinearRing({coordinates});
-}
-
-function toMultiPolygonGeometry(positions: Position[][][]): Pb.Geometry {
-  // Skip invalid polygons.
-  const polygons = positions.map(p => toPolygonPb(p));
-  if (polygons.length === 0) throw new Error('Empty multi-polygon');
-  const multiPolygon = new Pb.MultiPolygon({polygons});
-  return new Pb.Geometry({multiPolygon});
 }
