+++ lib/src/index.ts
@@ -17,5 +17,5 @@
 export {toDocumentData} from './proto-to-firestore';
 export {toMessage} from './firestore-to-proto';
 export {deleteEmpty, isEmpty} from './obj-util';
-export {toGeoJsonGeometry} from './geo-json';
+export {toGeoJsonGeometry, toGeometryPb} from './geo-json';
 export {registry} from './message-registry';
